requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-filter==2.4.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
kombu==4.6.11
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
vine==1.3.0
